// apps/core/src/core/theme/types.ts
export type RangeenThemeSchema = {
  background: string;
  foreground: string;
  card: string;
  cardForeground: string;
  popover: string;
  popoverForeground: string;
  primary: string;
  primaryForeground: string;
  secondary: string;
  secondaryForeground: string;
  muted: string;
  mutedForeground: string;
  accent: string;
  accentForeground: string;
  destructive: string;
  destructiveForeground: string;
  border: string;
  input: string;
  ring: string;
  radius: string;
};

export type RangeenTheme = {
  light: RangeenThemeSchema;
  dark: RangeenThemeSchema;
};